내 생각엔 level_03_02 끝 난 것 같음

+++ src/main/java/calculator/AddOperator.java
@@ -1,7 +1,7 @@
 package calculator;
 
 public class AddOperator {
-    public static double operate(int num1, int num2) {
+    public static double operate(double num1, double num2) {
         return num1 + num2;
     }
 }
