+++ src/main/java/calculator/ModOperator.java
@@ -2,8 +2,8 @@ package calculator;
 
 public class ModOperator {
     private static int mod;
-    public static int operate(int num1, int num2) {
-        setMod(num1 % num2);
+    public static int operate(double num1, double num2) {
+        setMod((int) (num1 % num2));
         return getMod();
     }
